Refactoring package structure: entity를 domain 패키지 하위로 위치 변경
