+++ src/main/java/study/tdd/simpleboard/api/post/domain/PageResponseDTO.java
@@ -2,11 +2,9 @@ package study.tdd.simpleboard.api.post.domain;
 
 import lombok.Getter;
 import org.springframework.data.domain.Page;
-import study.tdd.simpleboard.api.post.domain.PostPageDTO;
-import study.tdd.simpleboard.api.post.entity.Post;
+import study.tdd.simpleboard.api.post.domain.entity.Post;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Getter
 public class PageResponseDTO {
