+++ src/main/java/study/tdd/simpleboard/api/post/domain/enums/PostMessage.java
@@ -1,0 +1,4 @@
+package study.tdd.simpleboard.api.post.domain;
+
+public enum PostMessage {
+}
