+++ src/test/java/study/tdd/simpleboard/api/post/PostServiceTest.java
@@ -8,17 +8,14 @@ import org.springframework.data.domain.*;
 import study.tdd.simpleboard.api.member.entity.Member;
 import study.tdd.simpleboard.api.post.domain.PageResponseDTO;
 import study.tdd.simpleboard.api.post.domain.PostOneDTO;
-import study.tdd.simpleboard.api.post.domain.PostPageDTO;
-import study.tdd.simpleboard.api.post.entity.Post;
+import study.tdd.simpleboard.api.post.domain.entity.Post;
 import study.tdd.simpleboard.api.post.repository.PostRepository;
 import study.tdd.simpleboard.api.post.service.PostService;
 import study.tdd.simpleboard.exception.common.BizException;
 import study.tdd.simpleboard.exception.post.PostCrudErrorCode;
 
-import java.util.Iterator;
 import java.util.List;
 import java.util.Optional;
-import java.util.function.Function;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.BDDMockito.given;
